Use Exif for storing location coordinates
Uses Exif to store coordinates with the file instead of sending and storing it separately

+++ app/src/main/java/memphis/myapplication/UI/ViewPhotosFragment.java
@@ -5,6 +5,7 @@ import android.os.CountDownTimer;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.exifinterface.media.ExifInterface;
 import androidx.fragment.app.Fragment;
 import androidx.navigation.Navigation;
 
@@ -18,10 +19,12 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import com.squareup.picasso.Picasso;
 
 import java.io.File;
+import java.io.IOException;
 import java.util.ArrayList;
 
 /**
@@ -75,7 +78,7 @@ public class ViewPhotosFragment extends Fragment {
         new CountDownTimer(5050*photos.size(),5000) { // 5000 = 5 sec
 
             public void onTick(long tick) {
-                if(tick/5000>=1) {
+                if(tick/5000>=1 && isAdded() && getActivity() != null) {
                     String photo = photos.get(m_index);
                     File photoFile = new File(photo);
                     Picasso.get().load(photoFile).fit().centerCrop().into(m_imgView);
@@ -83,8 +86,45 @@ public class ViewPhotosFragment extends Fragment {
                     Timber.d("Displating: "+photo.substring(photo.lastIndexOf('_')+ 1));
                     FilesInfo filesInfo = RealmRepository.getInstance().getFileInfo(photo.substring(photo.lastIndexOf('_')+ 1));
                     if(filesInfo.location){
-                        m_location.setVisibility(View.VISIBLE);
-                        m_location.setText(filesInfo.latitude + " : "+filesInfo.longitude);
+                        float latitude = 0;
+                        float longitude = 0;
+                        ExifInterface exif;
+
+                        try {
+                            exif = new ExifInterface(photoFile.getAbsolutePath());
+                            String attrLATITUDE = exif.getAttribute(ExifInterface.TAG_GPS_LATITUDE);
+                            String attrLATITUDE_REF = exif.getAttribute(ExifInterface.TAG_GPS_LATITUDE_REF);
+                            String attrLONGITUDE = exif.getAttribute(ExifInterface.TAG_GPS_LONGITUDE);
+                            String attrLONGITUDE_REF = exif.getAttribute(ExifInterface.TAG_GPS_LONGITUDE_REF);
+                            Timber.d(attrLATITUDE + " : " + attrLONGITUDE );
+
+                            if((attrLATITUDE !=null) && (attrLATITUDE_REF !=null) && (attrLONGITUDE != null)
+                                    && (attrLONGITUDE_REF !=null))
+                            {
+                                if(attrLATITUDE_REF.equals("N")){
+                                    latitude = convertToDegree(attrLATITUDE);
+                                }
+                                else{
+                                    latitude = 0 - convertToDegree(attrLATITUDE);
+                                }
+
+                                if(attrLONGITUDE_REF.equals("E")){
+                                    longitude = convertToDegree(attrLONGITUDE);
+                                }
+                                else{
+                                    longitude = 0 - convertToDegree(attrLONGITUDE);
+                                }
+                            }
+                            
+                            Timber.d(latitude + " : " + longitude );
+                            m_location.setVisibility(View.VISIBLE);
+                            m_location.setText(latitude + " : "+ longitude);
+                        } catch (IOException e) {
+                            e.printStackTrace();
+                            m_location.setVisibility(View.GONE);
+                            m_location.setText("");
+                            Toast.makeText(getActivity(), "Error in getting location", Toast.LENGTH_SHORT).show();
+                        }
                     }else{
                         m_location.setVisibility(View.GONE);
                         m_location.setText("");
@@ -107,15 +147,39 @@ public class ViewPhotosFragment extends Fragment {
                 }
             }
 
-            public void onFinish() {
+        public void onFinish() {
                 // since we are deleting photos only after we set the next one, we need to delete the
                 // last photo
-                String photoToDelete = photos.get(m_index-1);
-                File fileToDelete = new File(photoToDelete);
-                fileToDelete.delete();
-                Navigation.findNavController(viewPhotosView).popBackStack();
+            if (isAdded() && getActivity() != null) {
+                    String photoToDelete = photos.get(m_index - 1);
+                    File fileToDelete = new File(photoToDelete);
+                    fileToDelete.delete();
+                    Navigation.findNavController(viewPhotosView).popBackStack();
+                }
             }
 
         }.start();
+    }
+    private static float convertToDegree(String stringDMS){
+        Float result;
+        String[] dms = stringDMS.split(",", 3);
+
+        String[] stringD = dms[0].split("/", 2);
+        double d0 = Double.valueOf(stringD[0]);
+        double d1 = Double.valueOf(stringD[1]);
+        double degree = d0/d1;
+
+        String[] stringM = dms[1].split("/", 2);
+        double m0 = Double.valueOf(stringM[0]);
+        double m1 = Double.valueOf(stringM[1]);
+        double minute = m0/m1;
+
+        String[] stringS = dms[2].split("/", 2);
+        double S0 = Double.valueOf(stringS[0]);
+        double S1 = Double.valueOf(stringS[1]);
+        double second = S0/S1;
+
+        result = (float) (degree + (minute / 60) + (second / 3600));
+        return result;
     }
 }
